+++ django_alt/abstract/serializers.py
@@ -1,5 +1,7 @@
 from collections import OrderedDict
 
+from django.core.exceptions import ObjectDoesNotExist
+
 from rest_framework import serializers
 from rest_framework.fields import empty
 
@@ -12,7 +14,6 @@ class BaseValidatedSerializer(serializers.Serializer):
     Defines a relation between rest_framework serializer and validator
     """
     FIELD_VALIDATOR_PREFIX = 'field_'
-    ATTRS_VALIDATOR_PREFIX = 'check_'
 
     def __init__(self, instance=None, data=empty, *, validator_class=None, **kwargs):
         if not hasattr(self, 'Meta'):
@@ -38,6 +39,15 @@ class BaseValidatedSerializer(serializers.Serializer):
         :return: None
         """
         return self.Meta.validator_class(serializer=self, **kwargs)
+
+    @staticmethod
+    def _check_permissions(permission_test, attrs):
+        try:
+            if permission_test is not True and permission_test is not None and not permission_test(attrs):
+                raise PermissionError()
+        except (ObjectDoesNotExist, KeyError):
+            raise PermissionError()
+        return True
 
     @property
     def validator(self) -> Validator:
@@ -66,7 +76,7 @@ class BaseValidatedSerializer(serializers.Serializer):
         as the parameter.
         If subclass defines functions with names starting with check_,
         executes such functions with attrs dict as the parameter.
-        All functions are called in alphabetical order
+        All functions are called in alphabetical order.
         :return: None
         """
         fields = sorted(self.fields.fields.keys())
@@ -75,10 +85,16 @@ class BaseValidatedSerializer(serializers.Serializer):
             if hasattr(self.validator, name) and callable(getattr(self.validator, name)) and field in attrs:
                 getattr(self.validator, name)(attrs[field])
 
-        def is_attr_action(name):
-            return name.startswith(self.ATTRS_VALIDATOR_PREFIX) and callable(getattr(self.validator, name))
-        for name in [name for name in dir(self.validator) if is_attr_action(name)]:
-            getattr(self.validator, name)(attrs)
+        self.validator.validate_checks(attrs)
+
+    def check_permissions(self, attrs):
+        """
+        Performs post-permission checking, if a `permission_test` callable
+        was passed to the serializer via its constructor
+        :return: None
+        :raises: PermissionError
+        """
+        self.permission_test is not None and self._check_permissions(self.permission_test, attrs)
 
     def validate(self, attrs: dict) -> dict:
         """
@@ -99,9 +115,8 @@ class BaseValidatedSerializer(serializers.Serializer):
 
         attrs = coal(self.validator.base_db(attrs), attrs)
 
-        # post-permission checking for other methods
-        if self.permission_test is not None and not self.permission_test(attrs):
-            raise PermissionError()
+        self.check_permissions(attrs)
+
         return attrs
 
     def to_representation(self, instance) -> OrderedDict: